+++ services/horizon/internal/db2/core/main_test.go
@@ -139,24 +139,6 @@ func TestSchemaVersion8(t *testing.T) {
 		connectedAsset := xdr.MustNewNativeAsset()
 		tt.Assert.True(assets[0].Equals(connectedAsset))
 	}
-
-	var orderbookSummary OrderBookSummary
-	err = q.GetOrderBookSummary(&orderbookSummary, xdr.MustNewNativeAsset(), xdr.MustNewCreditAsset("USD", "GB2QIYT2IAUFMRXKLSLLPRECC6OCOGJMADSPTRK7TGNT2SFR2YGWDARD"), 10)
-	if tt.Assert.NoError(err) {
-		checkOrderBookRow(tt, orderbookSummary[0], "ask", int32(1), int32(1), float64(1), "300000000")
-		checkOrderBookRow(tt, orderbookSummary[1], "ask", int32(3), int32(1), float64(3), "400000000")
-
-		checkOrderBookRow(tt, orderbookSummary[2], "bid", int32(1), int32(2), float64(0.5), "200000000")
-		checkOrderBookRow(tt, orderbookSummary[3], "bid", int32(1), int32(1), float64(1), "100000000")
-	}
-}
-
-func checkOrderBookRow(tt *test.T, row OrderBookSummaryPriceLevel, typ string, pricen, priced int32, pricef float64, amount string) {
-	tt.Assert.Equal(typ, row.Type)
-	tt.Assert.Equal(pricen, row.Pricen)
-	tt.Assert.Equal(priced, row.Priced)
-	tt.Assert.Equal(pricef, row.Pricef)
-	tt.Assert.Equal(amount, row.Amount)
 }
 
 func TestSchemaVersion9(t *testing.T) {
@@ -258,15 +240,5 @@ func TestSchemaVersion9(t *testing.T) {
 		tt.Assert.Equal(1, len(assets))
 		connectedAsset := xdr.MustNewNativeAsset()
 		tt.Assert.True(assets[0].Equals(connectedAsset))
-	}
-
-	var orderbookSummary OrderBookSummary
-	err = q.GetOrderBookSummary(&orderbookSummary, xdr.MustNewNativeAsset(), xdr.MustNewCreditAsset("USD", "GB2QIYT2IAUFMRXKLSLLPRECC6OCOGJMADSPTRK7TGNT2SFR2YGWDARD"), 10)
-	if tt.Assert.NoError(err) {
-		checkOrderBookRow(tt, orderbookSummary[0], "ask", int32(1), int32(1), float64(1), "300000000")
-		checkOrderBookRow(tt, orderbookSummary[1], "ask", int32(3), int32(1), float64(3), "400000000")
-
-		checkOrderBookRow(tt, orderbookSummary[2], "bid", int32(1), int32(2), float64(0.5), "200000000")
-		checkOrderBookRow(tt, orderbookSummary[3], "bid", int32(1), int32(1), float64(1), "100000000")
 	}
 }