+++ services/horizon/internal/db2/schema/bindata.go
@@ -28,6 +28,7 @@
 // migrations/33_remove_unused.sql (860B)
 // migrations/34_fee_bump_transactions.sql (863B)
 // migrations/35_drop_participant_id.sql (306B)
+// migrations/36_add_orderbook_index.sql (141B)
 // migrations/3_use_sequence_in_history_accounts.sql (447B)
 // migrations/4_add_protocol_version.sql (188B)
 // migrations/5_create_trades_table.sql (1.1kB)
@@ -664,6 +665,26 @@ func migrations35_drop_participant_idSql() (*asset, error) {
 	return a, nil
 }
 
+var _migrations36_add_orderbook_indexSql = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xd2\xd5\x55\xd0\xce\xcd\x4c\x2f\x4a\x2c\x49\x55\x08\x2d\xe0\xe2\x72\x0e\x72\x75\x0c\x71\x55\xf0\xf4\x73\x71\x8d\x50\x48\x4a\x2d\x2e\x89\xcf\x4f\x4b\x4b\x2d\x52\xf0\xf7\x53\x00\x33\x8a\x15\x42\x83\x3d\xfd\xdc\x15\x92\x4a\x8a\x52\x53\x15\x34\x8a\x53\x73\x72\x32\xf3\xd2\xe3\x13\x8b\x8b\x53\x4b\x74\x14\x92\x4a\x2b\x91\x78\x05\x45\x99\xc9\xa9\x9a\xd6\x5c\x5c\xc8\x96\xb8\xe4\x97\xe7\x71\x71\xb9\x04\xf9\x07\x60\x58\x62\x0d\x08\x00\x00\xff\xff\x2b\x0e\x55\x72\x8d\x00\x00\x00")
+
+func migrations36_add_orderbook_indexSqlBytes() ([]byte, error) {
+	return bindataRead(
+		_migrations36_add_orderbook_indexSql,
+		"migrations/36_add_orderbook_index.sql",
+	)
+}
+
+func migrations36_add_orderbook_indexSql() (*asset, error) {
+	bytes, err := migrations36_add_orderbook_indexSqlBytes()
+	if err != nil {
+		return nil, err
+	}
+
+	info := bindataFileInfo{name: "migrations/36_add_orderbook_index.sql", size: 0, mode: os.FileMode(0), modTime: time.Unix(0, 0)}
+	a := &asset{bytes: bytes, info: info, digest: [32]uint8{0xb5, 0xf1, 0x1b, 0x57, 0xa9, 0x94, 0xb, 0x9d, 0xa0, 0x9, 0x30, 0x7, 0x8, 0x9e, 0x51, 0xd3, 0x76, 0xed, 0x1e, 0x54, 0xa1, 0x22, 0x18, 0x3e, 0xf8, 0xe9, 0x6, 0xb8, 0xf8, 0x38, 0xea, 0xc5}}
+	return a, nil
+}
+
 var _migrations3_use_sequence_in_history_accountsSql = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\x9c\x91\x4d\x6b\xb3\x40\x14\x85\xf7\xf3\x2b\xce\x2e\xca\xfb\x66\x91\x6d\x5c\x4d\xc6\x1b\x22\x8c\x63\x3b\x5e\xdb\x64\x25\xa2\x43\x3a\x90\x6a\xeb\xd8\xaf\x7f\x5f\x48\xd3\x0f\x08\x6d\xa1\xcb\x73\x78\xe0\x39\xdc\x3b\x9f\xe3\xdf\xad\xdf\x8f\xcd\xe4\x50\xdd\x09\x65\x49\x32\xa1\xa4\xcb\x8a\x8c\x22\xdc\xf8\x30\x0d\xe3\x4b\xdd\xb4\xed\xf0\xd0\x4f\xa1\xf6\x5d\x1d\xdc\xbd\x00\x80\x92\xa5\x65\x5c\x67\xbc\xc1\xe2\x58\x64\x46\x59\xca\xc9\x30\x56\xbb\x53\x65\x0a\xe4\x99\xb9\x92\xba\xa2\x8f\x2c\xb7\x9f\x59\x49\xb5\x21\x2c\x12\x51\x92\x26\xc5\x08\x6e\x7a\x6c\x0e\xd1\xec\x1b\xef\xec\x3f\xa2\x13\x99\xcb\x6d\xe4\xbb\x18\x6b\x5b\xe4\x67\x33\xe3\x38\x11\x52\x33\x59\xb0\x5c\x69\x42\x61\xf4\xee\x0c\xc2\x1b\xa1\x0a\x5d\xe5\x06\xbe\x43\x49\x8c\x94\xd6\xb2\xd2\x8c\xde\x3d\xff\xbc\x64\xb9\x1c\xdd\xbe\x3d\x34\x21\xc4\x89\x10\x5f\xcf\x98\x0e\x4f\xfd\x1f\xec\xa9\x2d\x2e\xde\xf5\x89\x38\xa6\xdf\xde\x90\x88\xd7\x00\x00\x00\xff\xff\x55\xe2\xdd\x2c\xbf\x01\x00\x00")
 
 func migrations3_use_sequence_in_history_accountsSqlBytes() ([]byte, error) {
@@ -943,6 +964,7 @@ var _bindata = map[string]func() (*asset, error){
 	"migrations/33_remove_unused.sql":                         migrations33_remove_unusedSql,
 	"migrations/34_fee_bump_transactions.sql":                 migrations34_fee_bump_transactionsSql,
 	"migrations/35_drop_participant_id.sql":                   migrations35_drop_participant_idSql,
+	"migrations/36_add_orderbook_index.sql":                   migrations36_add_orderbook_indexSql,
 	"migrations/3_use_sequence_in_history_accounts.sql":       migrations3_use_sequence_in_history_accountsSql,
 	"migrations/4_add_protocol_version.sql":                   migrations4_add_protocol_versionSql,
 	"migrations/5_create_trades_table.sql":                    migrations5_create_trades_tableSql,
@@ -1023,6 +1045,7 @@ var _bintree = &bintree{nil, map[string]*bintree{
 		"33_remove_unused.sql":                         &bintree{migrations33_remove_unusedSql, map[string]*bintree{}},
 		"34_fee_bump_transactions.sql":                 &bintree{migrations34_fee_bump_transactionsSql, map[string]*bintree{}},
 		"35_drop_participant_id.sql":                   &bintree{migrations35_drop_participant_idSql, map[string]*bintree{}},
+		"36_add_orderbook_index.sql":                   &bintree{migrations36_add_orderbook_indexSql, map[string]*bintree{}},
 		"3_use_sequence_in_history_accounts.sql":       &bintree{migrations3_use_sequence_in_history_accountsSql, map[string]*bintree{}},
 		"4_add_protocol_version.sql":                   &bintree{migrations4_add_protocol_versionSql, map[string]*bintree{}},
 		"5_create_trades_table.sql":                    &bintree{migrations5_create_trades_tableSql, map[string]*bintree{}},