+++ services/horizon/internal/db2/schema/migrations/36_add_orderbook_index.sql
@@ -1,0 +1,7 @@
+-- +migrate Up
+
+CREATE INDEX best_offer ON offers USING btree (selling_asset, buying_asset, price);
+
+-- +migrate Down
+
+DROP INDEX best_offer;