+++ services/horizon/internal/web.go
@@ -210,9 +210,7 @@ func (w *web) mustInstallActions(config Config, pathFinder paths.Finder, orderBo
 				"/order_book",
 				streamableObjectActionHandler{
 					streamHandler: streamHandler,
-					action: actions.GetOrderbookHandler{
-						OrderBookGraph: orderBookGraph,
-					},
+					action:        actions.GetOrderbookHandler{},
 				},
 			)
 		}