+++ services/horizon/CHANGELOG.md
@@ -8,6 +8,7 @@ file. This project adheres to [Semantic Versioning](http://semver.org/).
 * Drop support for MuxedAccounts strkeys (spec'ed in [SEP23](https://github.com/stellar/stellar-protocol/blob/master/ecosystem/sep-0023.md)).
   SEP23 is still a draft and we don't want to encourage storing strkeys which may not be definite.
 * Replace `SequenceProvider` implementation with one which queries the Horizon DB for sequence numbers instead of the Stellar Core DB.
+* Use the Horizon DB instead of Horizon's in memory order book graph to query orderbook details for the /order_book endpoint.
 * Remove JSON variant of `GET /metrics`, both in the server and client code. It's using Prometheus format by default now.
 * Decreased a memory usage of initial state ingestion stage and state verifier ([#2618](https://github.com/stellar/go/pull/2618)).